Move manylinux logic into private module
