+++ tests/manylinux/hello-world-invalid-class
