+++ tests/manylinux/hello-world-invalid-magic
