+++ tests/manylinux/hello-world-x86_64-x32
