+++ tests/test_musllinux.py
@@ -33,6 +33,12 @@ LD_MUSL_I386 = "/lib/ld-musl-i386.so.1"
 LD_MUSL_AARCH64 = "/lib/ld-musl-aarch64.so.1"
 
 
+@pytest.fixture(autouse=True)
+def clear_lru_cache():
+    yield
+    _get_musl_version.cache_clear()
+
+
 @pytest.mark.parametrize(
     "output, version",
     [
